EditProfilePopup

+++ src/utils/api.js
@@ -71,26 +71,49 @@ class Api {
     .then(this._handleResponse)
   }
 
-  addLike(cardId) {
-    return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
-      method: 'PUT',
-      headers: this._headers,
-      body: JSON.stringify({
-        _id: cardId,
-      })
-    })
-    .then(this._handleResponse)
-  }
+  // addLike(cardId) {
+  //   return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+  //     method: 'PUT',
+  //     headers: this._headers,
+  //     body: JSON.stringify({
+  //       _id: cardId,
+  //     })
+  //   })
+  //   .then(this._handleResponse)
+  // }
 
-  delLike(cardId) {
-    return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
-      method: 'DELETE',
-      headers: this._headers,
-      body: JSON.stringify({
-        _id: cardId,
+  // delLike(cardId) {
+  //   return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+  //     method: 'DELETE',
+  //     headers: this._headers,
+  //     body: JSON.stringify({
+  //       _id: cardId,
+  //     })
+  //   })
+  //   .then(this._handleResponse)
+  // }
+
+
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+        method: 'PUT',
+        headers: this._headers,
+        body: JSON.stringify({
+          _id: cardId,
+        })
       })
-    })
-    .then(this._handleResponse)
+      .then(this._handleResponse)
+    } else {
+      return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+        method: 'DELETE',
+        headers: this._headers,
+        body: JSON.stringify({
+          _id: cardId,
+        })
+      })
+      .then(this._handleResponse)
+    }
   }
 }
